Add support for OpenCV 4.0, fix x64 build warnings, fix latex file generation, add 2019 copyright, change tabs to spaces

+++ include/memory_tester.h
@@ -1,4 +1,4 @@
-// Copyright(c) 2016 - 2018 Federico Bolelli, Costantino Grana, Michele Cancilla, Lorenzo Baraldi and Roberto Vezzani
+// Copyright(c) 2016 - 2019 Federico Bolelli, Costantino Grana, Michele Cancilla, Lorenzo Baraldi and Roberto Vezzani
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
@@ -69,7 +69,7 @@ public:
         cols = size.width;
     }
 
-    MemMat(size_t rows, size_t cols) : MemMat(cv::Size(cols, rows)) {}
+    MemMat(unsigned rows, unsigned cols) : MemMat(cv::Size(cols, rows)) {}
 
     MemMat(cv::Size size, const T val)
     {
@@ -79,7 +79,7 @@ public:
         cols = size.width;
     }
 
-    MemMat(size_t rows, size_t cols, const T val) : MemMat(cv::Size(cols, rows), val) {}
+    MemMat(unsigned rows, unsigned cols, const T val) : MemMat(cv::Size(cols, rows), val) {}
 
     T& operator()(const int r, const int c)
     {
